tests, binary sequences, subsets, etc.

+++ src/main/kotlin/math/divides.kt
@@ -1,5 +1,5 @@
 package math
 
-fun Long.divides(number: Long): Boolean {
-    return number % this == 0L
+fun Int.divides(number: Int): Boolean {
+    return number % this == 0
 }
